// app/modules/skills/types/skills.types.ts
export type SkillCategory =
  | "Frontend"
  | "Backend"
  | "Database"
  | "Mobile"
  | "API Integration"
  | "Cloud"
  | "DevOps"
  | "Tools"
  | "Other";